Support results 0.14 as a peer dependency
It should have been a peer dependency before, so now it is.

It's important that react-results-proptypes is using the same copy of `results` that your project is. Through some luck, we were getting react-results-proptypes' copy of results deduped by webpack, so stuff always worked.

As a peer dependency, we'll be forced to keep the version of `results` and `react-results-proptypes` in sync by npm. woo!

Bumping a minor, since this can break builds etc.